pyproject.toml
[project]
name = "investment-memo-orchestrator"
version = "0.1.0"
description = "Multi-agent orchestration system for generating investment memos"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",  # For nice terminal output
    "tavily-python>=0.3.0",  # Web search for research
    "httpx>=0.27.0",  # For web fetching
    "beautifulsoup4>=4.12.0",  # For HTML parsing
    "openai>=1.0.0",  # Optional: for Perplexity API (uses OpenAI SDK)
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
